Fix warnings.
git-svn-id: https://fuse-emulator.svn.sourceforge.net/svnroot/fuse-emulator/trunk@3334 6c7d1b9a-d430-0410-8293-a2b96dacb39b

+++ libspectrum/szx.c
@@ -489,7 +489,7 @@ read_plsd_chunk( libspectrum_snap *snap, libspectrum_word version GCC_UNUSED,
                                "in compressed file, should be %lu, file "
                                "has %lu",
 			       __FILE__, 
-                               expected_length,
+                               (unsigned long)expected_length,
                                (unsigned long)uncompressed_length );
       return LIBSPECTRUM_ERROR_UNKNOWN;
     }
@@ -507,7 +507,7 @@ read_plsd_chunk( libspectrum_snap *snap, libspectrum_word version GCC_UNUSED,
                                  "in compressed file, should be %lu, file "
                                  "has %lu",
                                  __FILE__, 
-                                 expected_length,
+                                 (unsigned long)expected_length,
                                  (unsigned long)disc_rom_length );
         return LIBSPECTRUM_ERROR_UNKNOWN;
       }
@@ -534,7 +534,7 @@ read_plsd_chunk( libspectrum_snap *snap, libspectrum_word version GCC_UNUSED,
                                "in uncompressed file, should be %lu, file "
                                "has %lu",
 			       __FILE__, 
-                               expected_length,
+                               (unsigned long)expected_length,
                                (unsigned long)disc_rom_length );
       return LIBSPECTRUM_ERROR_UNKNOWN;
     }
